* Warning fixes * Fixed cmakelists for algorithms
git-svn-id: file:///home/michael/svn/yuni/trunk@1987 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/tools/yuni-docmake/make/job-writer.cpp
@@ -78,7 +78,7 @@ namespace // anonymous
 					<< "\t\t\tlabel: \"" << curve.caption << "\",\n"
 					<< "\t\t\tdata: [";
 
-				unsigned int max = Math::Max(curve.x.size(), curve.y.size());
+				unsigned int max = static_cast<unsigned int>(Math::Max(curve.x.size(), curve.y.size()));
 				for (unsigned int pt = 0; pt != max; ++pt)
 				{
 					if (pt)
